// src/app/private-app/interfaces/info-evaluacion-representante.ts
export interface InfoEvaluacionRepresentante {
  razon_social: string;
  representante_legal: string;
  area_dedicacion: string;
  representante: string;
  nombre_estudiante: string;
  escuela: string;
  nivel: string;
  area_practicas: string;
  horas_practicas: number;
  fecha_inicio: string;
  fecha_fin: string;
}